requirements.txt
absl-py==1.4.0
aioice==0.9.0
aiortc==1.5.0
altair==4.2.2
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
av==10.0.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
cryptography==40.0.2
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
executing==1.2.0
flatbuffers==23.5.9
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==0.4.6
google-crc32c==1.5.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mtcnn==0.1.1
nest-asyncio==1.5.6
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
pyee==9.1.0
Pygments==2.15.1
pylibsrtp==0.8.0
Pympler==1.0.1
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.22.0
streamlit-webrtc==0.45.0
tenacity==8.2.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.15.0
get-ice-servers
streamlit-webrtc
